Added no cache headers, and moved back to an index.html for main webpage
Added no caching headers to the RESTful API output.  Might want to force 'cache' headers for organizations...not sure how ugly that'd look.  I've also removed the hardcoding of the username in index.php and change the /api/user/:id route to just /api/user and I'm using the logged in userId to populate said data.