Bug Fixes
Needed to change the order of the /log routes to allow /log/downloads to be accessed (needs to be first).  Also changed the Log class constructor to take an input in of OrgId.  This is used to create any necessesary checkin tables.  Aka instead of checking that the table exists for each method, we simply check on its construction.  Organization::createCheckinTable() will throw an error if invalid orgId, and it will create a checkin table if one does not already exist